requirements.txt
aiogram>=3.4.1,<4.0.0
openai>=1.30.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0

pypdf>=4.0.0,<5.0.0
python-docx>=1.1.0,<2.0.0
openpyxl>=3.1.2,<4.0.0
Pillow>=10.0.0,<12.0.0